Addresses non Ascii-7 characters in address comment by making guesses about encoding.
Generally speaking any char > 127 should result in a mimeEncoded word for the address comment
It is possible that some email clients don't read the RFC and will include NonAscii 7 characters

Because the characters 128-255 can vary according to string encoding and
there is no indication of string encoding in the header we have to guess.

The parser currently guesses in the following order ISO-Latin-1, ISO-Latin-2, Windows-Latin-1, Windows-Latin-2

This may result in the incorrect guess - but it is hard to correctly detect character set especially on small strings
such as address comments.